1. 修复下载文件，文件名乱码问题 2. 上传文件功能待修复

+++ request/HTTPRequest.h
@@ -46,7 +46,7 @@ public:
 
 private:
     string m_ip;
-    int m_port;
+    short m_port;
     bool m_open;
     SOCKET m_socket;
 };
